CI(pytest): Extract ansible mock in module
