+++ tests/lib/mock.py
@@ -184,28 +184,3 @@ def get_cvp_client(cvp_database) -> MagicMock:
     mock_client.api.get_images.side_effect = cvp_database.get_images
     mock_client.api.get_image_bundles.side_effect = cvp_database.get_image_bundles
     return mock_client
-
-
-class AnsibleFailJson(Exception):
-    """Exception class to be raised by module.fail_json and caught by the test case"""
-    pass
-
-
-def fail_json(msg: str = None):
-    raise AnsibleFailJson(msg)
-
-
-def get_ansible_module(check_mode: bool = False):
-    """
-    Return a mock ansible.module_utils.basic.AnsibleModule instance.
-    The test case could eventually verify that the module exited correcty by calling the `module.exit_json.assert_called()` method.
-
-    Returns
-    -------
-    MagicMock
-        The mock cpvrac.cvp_client.CvpClient instance.
-    """
-    mock_module = create_autospec(AnsibleModule)
-    mock_module.fail_json.side_effect = fail_json
-    mock_module.check_mode = check_mode
-    return mock_module
